Fix longstanding bug with State.traverse
Turns out state monads in strict languages are really tricky.
We can create a stack-safe foldlM relatively easily, but for traverse
we really need foldrM.

It seems that foldrM is not simply foldlM on the reversed list, so
traverse needs something else. For now I've settled on

    State.map List.reverse << State.foldl go []

It's stack-safe and passes the tests. If anyone is reading along here
and has other ideas, please open an issue.

Also exposes the tools for stack-safe operations in State